requirements.txt
asgiref==3.2.3
astroid==2.4.2
autopep8==1.4.4
beautifulsoup4==4.9.2
boto3==1.12.29
botocore==1.15.29
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.8
django-allauth==0.41.0
django-heroku==0.3.1
django-storages==1.9.1
django-tinymce==3.2.0
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
install==1.3.4
isort==5.6.4
jmespath==0.9.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pep8==1.7.1
Pillow==7.0.0
psycopg2==2.8.4
pycodestyle==2.5.0
pylint==2.6.0
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2018.5
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.14.0
soupsieve==2.0.1
sqlparse==0.2.4
toml==0.10.2
typed-ast==1.4.1
urllib3==1.25.8
whitenoise==5.0.1
wrapt==1.12.1
